+++ examples/area-chart.rs
@@ -1,12 +1,11 @@
 use plotters::prelude::*;
 
-use rand::distributions::Distribution;
-use rand::distributions::Normal;
 use rand::thread_rng;
+use rand_distr::{Distribution, Normal};
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let data: Vec<_> = {
-        let norm_dist = Normal::new(500.0, 100.0);
+        let norm_dist = Normal::new(500.0, 100.0).unwrap();
         let mut x_rand = thread_rng();
         let x_iter = norm_dist.sample_iter(&mut x_rand);
         x_iter