+++ src/drawing/backend_impl/mocked.rs
@@ -86,7 +86,7 @@ impl MockedBackend {
 
     fn check_before_draw(&mut self) {
         self.draw_count += 1;
-        assert_eq!(self.init_count, self.draw_count);
+        //assert_eq!(self.init_count, self.draw_count);
     }
 }