+++ src/drawing/backend_impl/svg.rs
@@ -1,6 +1,7 @@
 /*!
 The SVG image drawing backend
 */
+pub use svg as svg_types;
 
 use svg::node::element::{Circle, Line, Polygon, Polyline, Rectangle, Text};
 use svg::Document;
@@ -34,7 +35,7 @@ pub struct SVGBackend<'a> {
 }
 
 impl<'a> SVGBackend<'a> {
-    fn update_document<F: FnOnce(Document) -> Document>(&mut self, op: F) {
+    pub fn update_document<F: FnOnce(Document) -> Document>(&mut self, op: F) {
         let mut temp = None;
         std::mem::swap(&mut temp, &mut self.document);
         self.document = Some(op(temp.unwrap()));