+++ src/lib.rs
@@ -15,6 +15,10 @@
     <img style="display: inline! important" src="https://travis-ci.org/38/plotters.svg?branch=master"></img>
 </a>
 
+<a href="https://codecov.io/gh/38/plotters">
+  <img src="https://codecov.io/gh/38/plotters/branch/master/graph/badge.svg" />
+</a>
+
 Plotters is drawing library designed for rendering figures, plots, and charts, in pure rust. Plotters supports various types of backends,
 including bitmap, vector graph, piston window and WebAssembly.
 
@@ -219,7 +223,7 @@ including bitmap, vector graph, piston window and WebAssembly.
 To use Plotters, you can simply add Plotters into your `Cargo.toml`
 ```toml
 [dependencies]
-plotters = "^0.2.6"
+plotters = "^0.2.7"
 ```
 
 And the following code draws a quadratic function. `src/main.rs`,
@@ -630,7 +634,7 @@ pub mod prelude {
     };
 
     #[cfg(feature = "chrono")]
-    pub use crate::coord::{RangedDate, RangedDateTime};
+    pub use crate::coord::{RangedDate, RangedDateTime, RangedDuration};
 
     #[cfg(feature = "make_partial_axis")]
     pub use crate::coord::make_partial_axis;