+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-## Plotter latest (?)
+## Plotters latest (?)
+
+### Bug Fix
+
+- `FontError` from rusttype isn't `Sync` and `Send`. We don't have trait bound to ensure this.  (Thanks to @dalance)
+
+## Plotters latest (2019-10-1)
 
 ### Added